requirements.txt
fastapi
gunicorn
uvicorn
python-telegram-bot==20.3
python-dotenv
